+++ src/main/java/stonks/storage/Storage.java
@@ -1,8 +1,11 @@
+package stonks.storage;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import stonks.task.*;
 
 public class Storage {
     String filepath;
