+++ src/main/java/stonks/task/Deadline.java
@@ -1,3 +1,5 @@
+package stonks.task;
+
 public class Deadline extends Task {
     protected String deadline;
