+++ src/main/java/stonks/task/Event.java
@@ -1,3 +1,5 @@
+package stonks.task;
+
 public class Event extends Task {
     private String start;
     private String end;
