+++ src/main/java/stonks/task/Task.java
@@ -1,3 +1,5 @@
+package stonks.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
