+++ src/main/java/stonks/task/Todo.java
@@ -1,3 +1,5 @@
+package stonks.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
