+++ text-ui-test/data/stonks.txt
@@ -1,10 +1,2 @@
 T | 0 | borrow book
-T | 0 | borrow book
-D | 0 | return book | Sunday
-E | 0 | project meeting | Mon 2pm | 4pm
-T | 0 | borrow book
-D | 0 | return book | Sunday
-E | 0 | project meeting | Mon 2pm | 4pm
-T | 0 | borrow book
-D | 0 | return book | Sunday
 E | 0 | project meeting | Mon 2pm | 4pm
